added filter by keyword
